fix: change db & table name from constant

+++ app/src/main/java/com/exfarnanda1945/loremipsumgen/feat_history/data/entities/HistoryAndSetting.kt
@@ -70,7 +70,7 @@ fun HistoryGenerator.toEntity(): HistoryAndSetting {
             isReturnPlainText = setting.isReturnPlainText,
             numOfParagraphs = setting.numOfParagraphs,
             paragraphLength = setting.paragraphLength,
-            historyId = this.id.toLong()
+            historyId = this.id
         )
     )
 }
